% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_data}
\alias{get_data}
\title{Get data for a given GUID}
\usage{
get_data(base_url, api_token, guid)
}
\arguments{
\item{base_url}{The base url of the Junar service}

\item{api_token}{The user's API token for the Junar service}

\item{guid}{The user's API token for the Junar service}
}
\description{
Get the data for any given GUID
}
\keyword{GUID}

