% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_guid}
\alias{get_guid}
\title{Get a list of GUID's}
\usage{
get_guid(base_url, api_token)
}
\arguments{
\item{base_url}{The base url of the Junar service}

\item{api_token}{The user's API token for the Junar service}
}
\description{
Get a list of all the available GUID's with datasets or views from the
base URL
}
\keyword{GUID}

