% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_index}
\alias{get_index}
\title{Get a list of GUID's}
\usage{
get_index(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Get a list of all the available GUID's with datasets or views from the
base URL
}
\keyword{GUID}

