% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstats.R
\name{get_statsDS_Date}
\alias{get_statsDS_Date}
\title{Get hits by date}
\usage{
get_statsDS_Date(base_url, api_key, from, to)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Get a list of hits by date (hits to the datastream_date).
}
