% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstats.R
\name{get_statsDS_Month}
\alias{get_statsDS_Month}
\title{Get hits by groupMonth}
\usage{
get_statsDS_Month(base_url, api_key, month)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}

\item{month}{FIXME}
}
\description{
Get a list of hits by groupMonth
}
