% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstats.R
\name{get_statsVZ}
\alias{get_statsVZ}
\title{Get hits to visualizations}
\usage{
get_statsVZ(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Get a list of hits to visualizations
}
