% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{guid_list}
\alias{guid_list}
\title{Show GUID list}
\usage{
guid_list(base_url, api_token)
}
\arguments{
\item{base_url}{The base url of the Junar service}

\item{api_token}{The user's API token for the Junar service}
}
\description{
Show a simple list of all available GUID. This is not adding the titles to
the list because it generaly will not fit in the console window.
}
\keyword{GUID}

