% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{guid_titles}
\alias{guid_titles}
\title{Show GUID titles}
\usage{
guid_titles(base_url, api_token)
}
\arguments{
\item{base_url}{The base url of the Junar service}

\item{api_token}{The user's API token for the Junar service}
}
\description{
Show a simple list of all available GUID titles. This is not adding the
titles to the list because it generaly will not fit in the console window.
}
\keyword{GUID}

