% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{list_guid}
\alias{list_guid}
\title{Show GUID list}
\usage{
list_guid(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Show a list of all available GUID's identifying the available data sets. It
only shows GUID's so that the list will fit in the console window.
}
\keyword{GUID}
