% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{list_titles}
\alias{list_titles}
\title{Show GUID titles}
\usage{
list_titles(base_url, api_key)
}
\arguments{
\item{base_url}{The base url of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Show a simple list of all available GUID titles. This is not adding the
titles to the list because it generaly will not fit in the console window.
}
\keyword{GUID}

