% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{list_titles}
\alias{list_titles}
\title{Show GUID titles}
\usage{
list_titles(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL of the Junar service}

\item{api_key}{The user's API key for the Junar service}
}
\description{
Show a list of all available GUID titles. It only shows the titles so that
the list will fit in the console window.
}
\keyword{GUID}

